Move unit test of protobuff tool to component.
