+++ misc/protobuf-ballerina/src/test/java/org/ballerinalang/protobuf/util/BTestUtils.java
@@ -14,7 +14,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.test.net.grpc.util;
+package org.ballerinalang.protobuf.util;
 
 import org.ballerinalang.compiler.CompilerPhase;
 import org.ballerinalang.launcher.LauncherUtils;
