+++ misc/protobuf-ballerina/src/test/resources/protoFiles/helloWorld.proto
