+++ misc/protobuf-ballerina/src/test/resources/protoFiles/helloWorldClientStreaming.proto
